# Avoid costly Message.GetTargetHistory() on gateway reroute