--- a/src/Orleans.Runtime/Core/Dispatcher.cs
+++ b/src/Orleans.Runtime/Core/Dispatcher.cs
@@ -569,15 +569,6 @@ namespace Orleans.Runtime
             }
         }
 
-        /// <summary>
-        /// Reroute a message coming in through a gateway
-        /// </summary>
-        /// <param name="message"></param>
-        internal void RerouteMessage(Message message)
-        {
-            ResendMessageImpl(message);
-        }
-
         internal bool TryResendMessage(Message message)
         {
             if (!message.MayResend(this.messagingOptions.MaxResendCount)) return false;
@@ -601,7 +592,10 @@ namespace Orleans.Runtime
         private void ResendMessageImpl(Message message, ActivationAddress forwardingAddress = null)
         {
             if (logger.IsEnabled(LogLevel.Debug)) logger.Debug("Resend {0}", message);
-            message.TargetHistory = message.GetTargetHistory();
+            if (message.ForwardCount > 1)
+            {
+                message.TargetHistory = message.GetTargetHistory();
+            }
 
             if (message.TargetGrain.IsSystemTarget)
             {